Add typewriter effect to tooling
